requirements.txt
aerich==0.8.1
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.1.0
bcrypt==4.2.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dictdiffer==0.9.0
fastadmin==0.2.19
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.5
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pypika-tortoise==0.5.0
python-decouple==3.8
pytz==2025.1
redis==5.2.1
sniffio==1.3.1
starlette==0.45.3
tortoise-orm==0.24.0
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
